fixed compile_commands.json
since the file was generated on the container, the path did not reflect
the location of the repository on the host. added a script in the tools/
directory that is invoked whenever `make configure` is run which fixes
the paths used in compile_commands.json to point to the right files on
the host. for files that are exclusive to the container, they get copied
over to the build/ directory, for header files that are already
available on the host clang install, it points there. for those that are
not available on the host clang install, it points to the copied headers
in build/ti-headers